Initial code documentation

+++ include/util.h
@@ -14,8 +14,8 @@ void normal_exit(void);
 void terminate(unsigned char *str);
 void terminate_emsg(int emsg, unsigned char *str2);
 void alloc_bytes(unsigned char **buf, long bytes, unsigned char *id);
-void alloc_words(int **buf, long words, unsigned char *id);
-void alloc_dwords(long **buf, long dwords, unsigned char *id);
+void alloc_words(int16_t **buf, long words, unsigned char *id);
+void alloc_dwords(int32_t **buf, long dwords, unsigned char *id);
 void get_offset_length(int db_rec, long *offset, long *length);
 void get_offset(int db_rec, long *offset);
